Move files, Make border optional
